moved pages to their own files and exported from index files
