+++ src/modules/nominees/types.ts
@@ -1,0 +1,4 @@
+export type Nominee = {
+  id: string;
+  votes: number;
+};
