+++ src/modules/winners/types.ts
@@ -1,0 +1,4 @@
+export type Winner = {
+  id: string;
+  term: string;
+};
