+++ src/pages/auth/Auth.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import Page from 'pages/page';
+
+const Auth = () => {
+  return <Page></Page>;
+};
+
+export default Auth;
