+++ src/pages/auth/index.tsx
@@ -1,8 +1,2 @@
-import React from 'react';
-import Page from 'pages/page';
-
-const Auth = () => {
-  return <Page></Page>;
-};
-
-export default Auth;
+export { default } from './Auth';
+export * from './Auth';
