+++ src/pages/events/index.tsx
@@ -1,15 +1,2 @@
-import React from 'react';
-import Page from 'pages/page';
-import AddEventsForm from 'components/forms/AddEventsForm';
-import EventsList from 'components/EventsList';
-
-const Events = () => {
-  return (
-    <Page>
-      <EventsList></EventsList>
-      <AddEventsForm></AddEventsForm>
-    </Page>
-  );
-};
-
-export default Events;
+export { default } from './Events';
+export * from './Events';
