+++ src/pages/home/Home.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import Page from 'pages/page';
+
+const Home = () => {
+  return <Page></Page>;
+};
+
+export default Home;
