+++ src/pages/home/index.tsx
@@ -1,8 +1,2 @@
-import React from 'react';
-import Page from 'pages/page';
-
-const Home = () => {
-  return <Page></Page>;
-};
-
-export default Home;
+export { default } from './Home';
+export * from './Home';
