+++ src/pages/page/index.ts
@@ -1,0 +1,2 @@
+export { default } from './Page';
+export * from './Page';
