+++ src/pages/vote/index.ts
@@ -1,0 +1,2 @@
+export { default } from './Vote';
+export * from './Vote';
